Move update method to utils
